System working
set_zero and computation of values out of raw_values and offsets are done on server in glovecontroller class. only final values (float) are send to hololens. Debug.Log does delay server app

+++ Glove_Hololens_App/Assets/Code_Max/TrackingData.cs
@@ -12,13 +12,13 @@ public class TrackingData {
     public double timestamp;
     */
 
-    public long[] JointValues;
+    public float[] JointValues;
     public Matrix4x4 pose;
     public Vector3 velocity;
     public Vector3 acceleration;
     public double timestamp;
 
-    public TrackingData(long[] JointValues, Matrix4x4 pose, Vector3 velocity, Vector3 acceleration, double timestamp)
+    public TrackingData(float[] JointValues, Matrix4x4 pose, Vector3 velocity, Vector3 acceleration, double timestamp)
     {
         this.JointValues = JointValues;
         this.pose = pose;
